requirements.txt
fastai==2.8.2
huggingface_hub==0.33.0
numpy==2.3.0
Pillow==11.2.1
scikit-image==0.25.2
streamlit==1.45.1
timm==1.0.15
torch==2.7.1
torchvision==0.22.1
